requirements.txt
Django==2.2.4
django-tinymce==3.3.0
html5lib==1.1
Pillow==6.1.0
pisa==3.0.33
PyPDF2==1.26.0
reportlab==3.6.1
six==1.10.0
xhtml2pdf==0.2.5
yos-social-sdk==0.0.2
dj-static==0.0.6
static==1.1.1
whitenoise==5.2.0
wsgiref==0.1.2
psycopg2==2.8.6
